server/scripts_python/requirements.txt
Flask
pdfplumber
whisper
pymysql
requests
python-dotenv
scikit-learn
nltk
opencv-python
deepface
transformers
joblib
pandas
xgboost
numpy
